requirements.txt
imapclient
pyzmail36
html2text
python-dotenv
requests